LPS-55194 Move Wiki RSS settings to GroupService
