+++ modules/apps/wiki/wiki-api/src/com/liferay/wiki/settings/WikiGroupServiceSettingsOverride.java
@@ -27,4 +27,6 @@ public interface WikiGroupServiceSettingsOverride {
 
 	public String emailPageUpdatedSubjectXml();
 
+	public boolean enableRss();
+
 }
