+++ modules/apps/wiki/wiki-web/docroot/html/portlet/wiki/view_recent_changes.jsp
@@ -32,11 +32,11 @@ WikiNode node = (WikiNode)request.getAttribute(WikiWebKeys.WIKI_NODE);
 
 <br />
 
-<c:if test="<%= wikiPortletInstanceSettings.enableRss() %>">
+<c:if test="<%= wikiGroupServiceSettings.enableRss() %>">
 	<liferay-ui:rss
-		delta="<%= GetterUtil.getInteger(wikiPortletInstanceSettings.rssDelta()) %>"
-		displayStyle="<%= wikiPortletInstanceSettings.rssDisplayStyle() %>"
-		feedType="<%= wikiPortletInstanceSettings.rssFeedType() %>"
+		delta="<%= GetterUtil.getInteger(wikiGroupServiceSettings.rssDelta()) %>"
+		displayStyle="<%= wikiGroupServiceSettings.rssDisplayStyle() %>"
+		feedType="<%= wikiGroupServiceSettings.rssFeedType() %>"
 		url='<%= themeDisplay.getPathMain() + "/wiki/rss?p_l_id=" + plid + "&nodeId=" + node.getNodeId() %>'
 	/>
 </c:if>
